requirements.txt
discord.py
python-dotenv
asyncpraw
async-timeout 